Added derivative plot, and added many bugfixes for bugs from updating dependencies.
Derivative plot has very approximate scaling, attempted to get it to show the change in pressure in approximate kpa per day.

The old method of converting ordinal dates to pandas timestamps changed (matplotlib api change) so I completely rewrote that.
Also, the timezone adjusted date is saved in the .pkl file so it does not have to be re-applied after loading.

The api for using the data from pydap seems to have changed, and I just stumbled my way through fixing it.